fix(marketing): harden landing interactions
